[FIX] website_*: set correct order on modal footer button
Modal primary button should be the one in the corner (left if buttons aligned
left, right if buttons aligned right) as in the Bootstrap doc.
More than a simple guideline, the user will intuitively click on the button in
the corner as it is supposed to be the primary one.

Plus some DOM fixes (o_technical_modal was missing, H2 was too big especially
since BS4)

Closes #29890
task-1922120