[FIX] mrp: generate lot remove qty reserved/to consumer
commit 8eeb76c601c68eba865580079ec7a54dfb566198 introduced a button in
order to quickly add a lot number on the current production. However on
the produce wizard it will create the lot, save the wizard with the new
lot and reopen its view in edit mode. However the qty to consume and
reserved on the produce lines are readonly thus not saved and lost
during the process.

This commit add force-save on the qty to consume and reserved

closes odoo/odoo#29925