[IMP] mail: reflect blacklist inheritance on model
Like already done for some mail mixins

  * mail.thread reflected on ir.model: see c891e82fd48bfa40b723ed2ee286ab2505da83d7;
  * mail.activity.mixin reflected on ir.model: see 8aac8ba49bb761f95840dc52fb13c20006667b7f;

We allow to reflect the mail.blacklist.mixin on ir.model in this commit. This
inheritance imply the mail.address.mixin to allow normalization of emails on
models using the blacklist mixin.

It is useful notably to clearly store this inheritance and also allow
future tweaking through studio to enable more advanced mail capabilities
to custom models. Blacklist is useful when dealing with mass mailing and
marketing automation.

It will also be used in mailgateway to find all blacklist-related models
in Odoo and update the blacklist accordingly to the bounce and error
management in the mailgateway. Those improvements will come soon.

Related to task ID 1911679
Linked to PR #29483