[IMP] web: properly set margin for the kanban dropdown
While everything looks pretty perfect without this commit,
setting margin on `.dropdown-toggle` and not on its parent
`.o_dropdown_kanban` itself will cause glitches in Studio.

While selecting the dropdown using Studio, the class responsible
for displaying outline (that indicates the selection) is applied
on '.o_dropdown_kanban', which is okay, but the actual child is
positioned using margin, so the outline on parent div looks ugly.

This commit fixes the issue by applying the margin on the proper
selector.

Task 1916583

closes odoo/odoo#30049

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>